refactor Emotiv stuff into separate directory, add comments
